api/libraryquest_api/requirements.txt

Django==4.2.16
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
djoser==2.2.3
drf-spectacular==0.28.0
psycopg2-binary==2.9.9